# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Inventory Management and Audit System API",
        "description": "A comprehensive API for inventory management with audit logging, role-based access control, and real-time notifications",
        "contact": {
            "name": "API Support",
            "email": "admin@inventory-system.com"
        },
        "license": {
            "name": "MIT License",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Inventory Management API Server"
        },
        {
            "url": "http://localhost:8000/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/inventory-items": {
            "get": {
                "tags": [
                    "Inventory Items"
                ],
                "summary": "List inventory items",
                "description": "Returns a paginated list of inventory items with filtering options",
                "operationId": "inventoryItemIndex",
                "parameters": [
                    {
                        "name": "category",
                        "in": "query",
                        "description": "Filter by category",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "electronics",
                                "furniture",
                                "clothing",
                                "other"
                            ]
                        }
                    },
                    {
                        "name": "warehouse_id",
                        "in": "query",
                        "description": "Filter by warehouse ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "below_min_stock",
                        "in": "query",
                        "description": "Filter items with quantity below min_stock_level",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search in name or description",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/InventoryItemResource"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Inventory Items"
                ],
                "summary": "Create a new inventory item",
                "description": "Creates a new inventory item with image upload",
                "operationId": "inventoryItemStore",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Laptop XPS 15"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Dell XPS 15 laptop with 16GB RAM"
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "min_stock_level": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "unit_price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 1299.99
                                    },
                                    "category": {
                                        "type": "string",
                                        "enum": [
                                            "electronics",
                                            "furniture",
                                            "clothing",
                                            "other"
                                        ],
                                        "example": "electronics"
                                    },
                                    "warehouse_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Inventory item created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/InventoryItemResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/inventory-items/{id}": {
            "get": {
                "tags": [
                    "Inventory Items"
                ],
                "summary": "Get inventory item details",
                "description": "Returns details of a specific inventory item",
                "operationId": "inventoryItemShow",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Inventory item ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/InventoryItemResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Inventory item not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Inventory Items"
                ],
                "summary": "Update an inventory item",
                "description": "Updates an existing inventory item including stock adjustment",
                "operationId": "inventoryItemUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Inventory item ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Updated Laptop XPS 15"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Updated Dell XPS 15 laptop with 32GB RAM"
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "example": 15
                                    },
                                    "min_stock_level": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "unit_price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 1499.99
                                    },
                                    "category": {
                                        "type": "string",
                                        "enum": [
                                            "electronics",
                                            "furniture",
                                            "clothing",
                                            "other"
                                        ],
                                        "example": "electronics"
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Inventory item updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Inventory item updated successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Inventory item not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Inventory Items"
                ],
                "summary": "Delete an inventory item",
                "description": "Soft deletes an inventory item (admin only)",
                "operationId": "inventoryItemDestroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Inventory item ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Inventory item deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Inventory item deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden - Only admins can delete inventory items"
                    },
                    "404": {
                        "description": "Inventory item not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "InventoryItemResource": {
                "title": "Inventory Item Resource",
                "description": "Inventory item resource representation",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Laptop XPS 15"
                    },
                    "description": {
                        "type": "string",
                        "example": "Dell XPS 15 laptop with 16GB RAM"
                    },
                    "quantity": {
                        "type": "integer",
                        "example": 10
                    },
                    "min_stock_level": {
                        "type": "integer",
                        "example": 5
                    },
                    "unit_price": {
                        "type": "number",
                        "format": "float",
                        "example": 1299.99
                    },
                    "category": {
                        "type": "string",
                        "enum": [
                            "electronics",
                            "furniture",
                            "clothing",
                            "other"
                        ],
                        "example": "electronics"
                    },
                    "image": {
                        "type": "string",
                        "example": "inventory/2023-06-09/user_1_laptop.jpg"
                    },
                    "image_url": {
                        "type": "string",
                        "example": "http://localhost:8000/storage/inventory/2023-06-09/user_1_laptop.jpg"
                    },
                    "is_low_stock": {
                        "type": "boolean",
                        "example": false
                    },
                    "warehouse": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 1
                            },
                            "name": {
                                "type": "string",
                                "example": "Main Warehouse"
                            },
                            "location": {
                                "type": "string",
                                "example": "New York"
                            }
                        },
                        "type": "object"
                    },
                    "user": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 1
                            },
                            "name": {
                                "type": "string",
                                "example": "John Doe"
                            }
                        },
                        "type": "object"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-06-09T12:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-06-09T12:30:00Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}